add panel charts
